app.py
Senticore is a Streamlit-based web app that analyzes the sentiment of user-entered text using Natural Language Processing (NLP) and Machine Learning. It cleans and processes the text, then predicts whether the sentiment is Positive 😂, Neutral 😐, or Negative 👺. With a simple pastel-themed interface and instant predictions, Senticore is perfect for analyzing reviews, comments, or feedback in real time.